requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
dateparser==1.1.8
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
rest-framework-simplejwt==0.0.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
